Allow setting fractional com_maxfps.
This way one can specify 59.94 to match monitor refresh rate.
Although I'd say it does not make any difference =)


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10660 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/Common.cpp
@@ -2500,7 +2500,7 @@ void idCommonLocal::Frame( void ) {
 		static int64_t lastFrameAstroTime = Sys_Microseconds();
 		if (sessLocal.com_fixedTic.GetBool()) {
 			//stgatilov #4865: impose artificial FPS limit
-			int64_t minDeltaTime = 1000000 / sessLocal.com_maxFPS.GetInteger();
+			int64_t minDeltaTime = int64_t(1000000 / sessLocal.com_maxFPS.GetFloat());
 			int64_t currFrameAstroTime;
 			while (1) {
 				currFrameAstroTime = Sys_Microseconds();
